fix: clear gui followup quality regressions
